extracted EventStreamSource and EventStreamSink to interfaces
