--- src/SAREhub/Client/Event/EventStream.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace SAREhub\Client\Event;
-
-abstract class EventStream {
-	
-}
