+++ src/SAREhub/Client/Event/NullEventStreamSink.php
@@ -6,9 +6,17 @@ namespace SAREhub\Client\Event;
 /**
  * Class used as mock event stream sink
  */
-class NullEventStreamSink extends EventStreamSink {
+class NullEventStreamSink implements EventStreamSink {
 	
 	public function write(EventEnvelope $eventEnvelope) {
+		
+	}
+	
+	public function onPipe(EventStreamSource $source) {
+		
+	}
+	
+	public function onUnpipe(EventStreamSource $source) {
 		
 	}
 }
